# packages/contracts/artifacts/identity-workspace.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pertexo Identity and Workspace API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/auth/oidc/start": {
      "get": {
        "operationId": "startOidcLogin",
        "responses": {
          "200": {
            "description": "OIDC authorization transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OidcStartResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Unexpected"
          }
        }
      }
    },
    "/v1/auth/oidc/callback": {
      "get": {
        "operationId": "completeOidcLogin",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "minLength": 1,
              "maxLength": 4096
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "minLength": 16,
              "maxLength": 512
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Browser session established"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/Unexpected"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "logoutSession",
        "security": [
          {
            "cookieSession": []
          }
        ],
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "minLength": 16,
              "maxLength": 256
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Browser session revoked"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/Unexpected"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiProblem": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "readOnly": true,
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "instance": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048
          },
          "code": {
            "type": "string",
            "enum": [
              "auth.unauthenticated",
              "auth.forbidden",
              "resource.not_found",
              "request.invalid",
              "request.precondition_required",
              "request.idempotency_conflict",
              "request.rate_limited",
              "request.rate_limit_unavailable",
              "workspace.quota_exceeded",
              "artifact.conflict",
              "artifact.unavailable",
              "workspace.conflict",
              "workflow.revision_conflict",
              "workflow.lifecycle_conflict",
              "workflow.invalid",
              "workflow.not_published",
              "workflow.activation_failed",
              "run.not_cancelable",
              "run.outcome_unknown",
              "connection.conflict",
              "connection.reauthorization_required",
              "connection.revoked",
              "provider.rate_limited",
              "provider.unavailable",
              "platform.write_paused",
              "webhook.authentication_failed",
              "webhook.payload_too_large",
              "webhook.unsupported_media_type",
              "webhook.invalid_json",
              "webhook.idempotency_conflict",
              "webhook.rate_limited",
              "webhook.unavailable",
              "internal.unexpected"
            ]
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
          },
          "errors": {
            "readOnly": true,
            "maxItems": 100,
            "type": "array",
            "items": {
              "readOnly": true,
              "type": "object",
              "properties": {
                "path": {
                  "type": "string",
                  "maxLength": 1024
                },
                "code": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "message": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 500
                }
              },
              "required": [
                "path",
                "code",
                "message"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "requestId"
        ],
        "additionalProperties": false
      },
...
      "OidcStartResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "authorizationUrl": {
            "type": "string",
            "format": "uri"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "authorizationUrl",
          "expiresAt"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ApiProblem"
            }
          }
        }
      },
      "Unauthenticated": {
        "description": "Authentication required",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ApiProblem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ApiProblem"
            }
          }
        }
      },
...
      "ServiceUnavailable": {
        "description": "Upstream service unavailable",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ApiProblem"
            }
          }
        }
      },
...
      "Unexpected": {
        "description": "Unexpected server error",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ApiProblem"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "pertexo_session"
      }
    }
  }
}